+++ src/main/java/com/liquibaseproject/exception/ResultsNotFoundException.java
@@ -1,0 +1,7 @@
+package com.liquibaseproject.exception;
+
+public class ResultsNotFoundException extends RuntimeException {
+    public ResultsNotFoundException(String message) {
+        super(message);
+    }
+}